test/requirements.txt
pillow 
python-multipart 
requests
starlette
torch==2.0.1
torchvision==0.15.2
numpy
scikit-image